# ai-video-studio/codex/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Video Studio API",
    "version": "0.1.0",
    "description": "Semantic API contract for the Cutpilot/AI Video Studio MVP. Provider/model names are internal only."
  },
  "servers": [
    { "url": "https://api.example.local" }
  ],
  "paths": {
...
    "/shots/{shotId}/regenerate": {
      "post": {
        "operationId": "regenerate",
        "parameters": [
          { "name": "shotId", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^sht_" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["scope"],
                "additionalProperties": false,
                "properties": {
                  "scope": { "type": "string", "enum": ["shot", "segment"] },
                  "range": {
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                      "startSec": { "type": "number", "minimum": 0 },
                      "endSec": { "type": "number", "minimum": 0 }
                    }
                  },
                  "tweaks": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Regeneration queued",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "takes": { "type": "array", "items": { "$ref": "../schemas/domain.schema.json#/$defs/Take" } },
                    "jobs": { "type": "array", "items": { "$ref": "../schemas/domain.schema.json#/$defs/GenerationJob" } }
                  },
                  "required": ["takes", "jobs"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid regeneration request",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/ErrorResponse" }
              }
            }
          },
          "402": {
            "description": "Insufficient credits to reserve this work",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/InsufficientCreditsResponse" }
              }
            }
          },
          "404": {
            "description": "Shot not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/takes/{takeId}/upgrade": {
      "post": {
        "operationId": "upgradeTake",
        "parameters": [
          { "name": "takeId", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^tak_" } }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "mode": { "type": "string", "enum": ["auto", "final_regenerate", "enhance", "render_upscale"], "default": "auto" }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Upgrade queued",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "take": { "$ref": "../schemas/domain.schema.json#/$defs/Take" },
                    "job": { "$ref": "../schemas/domain.schema.json#/$defs/GenerationJob" }
                  },
                  "required": ["take", "job"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid upgrade request",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/ErrorResponse" }
              }
            }
          },
          "402": {
            "description": "Insufficient credits to reserve this work",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/InsufficientCreditsResponse" }
              }
            }
          },
          "404": {
            "description": "Done take not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/system/worker-leases/{leaseId}/renew": {
      "post": {
        "operationId": "renewWorkerLease",
        "parameters": [
          { "name": "leaseId", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^wlease_" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token"],
                "additionalProperties": false,
                "properties": {
                  "token": { "type": "string" },
                  "ttlSec": { "type": "integer", "minimum": 5, "maximum": 600 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Worker lease renewed",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/WorkerLeaseRenewResult" }
              }
            }
          },
          "404": {
            "description": "Lease not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/WorkerLeaseRenewResult" }
              }
            }
          },
          "409": {
            "description": "Lease token mismatch or lease is not active",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/WorkerLeaseRenewResult" }
              }
            }
          },
          "400": {
            "description": "Invalid worker lease renew request",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Admin access token required in production mode",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/ErrorResponse" }
              }
            }
          },
          "503": {
            "description": "Admin access token is not configured in production mode",
            "content": {
              "application/json": {
                "schema": { "$ref": "../schemas/domain.schema.json#/$defs/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  }
}